# medvision-cxr/frontend/public/openapi/medvision-cxr-openapi.yaml
openapi: 3.1.0
info:
  title: MedVision-CXR Backend
  version: 0.1.0
  summary: OpenAPI 3.1 specification for the MedVision-CXR backend
  description: MedVision-CXR is an AI-assisted chest X-ray triage system. This specification
    covers authentication, upload, AI-assisted risk assessment, Grad-CAM, doctor review,
    audit logs, model metadata, direct delete, health, and the governed deletion approval
    workflow.
paths:
...
  /api/v1/deletions/requests:
    post:
      tags:
      - deletions
      summary: Submit deletion request
      description: 提交胸片删除请求。普通用户仅可为自己上传的影像发起删除；doctor/admin 也可代表业务流程发起。
      operationId: submit_deletion_request_api_v1_deletions_requests_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeletionRequestCreate'
      responses:
        '201':
          description: Deletion request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionRequestOut'
        '400':
          description: Invalid delete mode or payload
        '403':
          description: User cannot request deletion for this image
        '404':
          description: Image not found
        '409':
          description: Pending deletion request already exists
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - deletions
      summary: List deletion requests
      description: 列出删除请求。doctor/admin 可查看全量请求，普通用户仅能查看自己发起的请求。
      operationId: get_deletion_requests_api_v1_deletions_requests_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: request_status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Status
      responses:
        '200':
          description: Deletion requests returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionRequestListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/deletions/requests/{deletion_request_id}/decision:
    post:
      tags:
      - deletions
      summary: Approve or reject deletion request
      description: 由 doctor/admin 审批删除请求。approve 会执行删除流程，reject 会保留拒绝原因与审批备注。
      operationId: review_deletion_request_api_v1_deletions_requests__deletion_request_id__decision_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: deletion_request_id
        in: path
        required: true
        schema:
          type: string
          title: Deletion Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeletionDecisionRequest'
      responses:
        '200':
          description: Deletion request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionRequestOut'
        '400':
          description: Invalid action or missing rejection_reason
        '403':
          description: Insufficient permissions
        '404':
          description: Deletion request or image not found
        '409':
          description: Deletion request already processed
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reviews:
    post:
      tags:
      - reviews
      summary: Create Review
      operationId: create_review_api_v1_reviews_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/reviews/{prediction_id}:
    get:
      tags:
      - reviews
      summary: Get Review
      operationId: get_review_api_v1_reviews__prediction_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: prediction_id
        in: path
        required: true
        schema:
          type: string
          title: Prediction Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DeletionDecisionRequest:
      properties:
        approval_action:
          type: string
          title: Approval Action
          description: approve or reject
        approval_note:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Approval Note
        rejection_reason:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Rejection Reason
      type: object
      required:
      - approval_action
      title: DeletionDecisionRequest
      examples:
      - approval_action: approve
        approval_note: 已核对病例状态，允许执行软删除。
      - approval_action: reject
        approval_note: 建议保留影像用于审计追踪。
        rejection_reason: 该影像已进入医生复核流程，当前不允许删除。
    DeletionRequestCreate:
      properties:
        image_id:
          type: string
          format: uuid
          title: Image Id
        delete_mode:
          type: string
          title: Delete Mode
          default: soft
        reason:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Reason
      type: object
      required:
      - image_id
      title: DeletionRequestCreate
      example:
        delete_mode: soft
        image_id: 11111111-1111-1111-1111-111111111111
        reason: 影像上传错误，请求撤回并重新上传。
    DeletionRequestListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/DeletionRequestOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - items
      - total
      title: DeletionRequestListResponse
    DeletionRequestOut:
      properties:
        request_id:
          type: string
          format: uuid
          title: Request Id
        image_id:
          type: string
          format: uuid
          title: Image Id
        delete_mode:
          type: string
          title: Delete Mode
        deletion_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Deletion Reason
        request_status:
          type: string
          title: Request Status
        requested_by:
          type: string
          format: uuid
          title: Requested By
        approved_by:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Approved By
        approval_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Approval Note
        rejection_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Rejection Reason
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - request_id
      - image_id
      - delete_mode
      - deletion_reason
      - request_status
      - requested_by
      - approved_by
      - approval_note
      - rejection_reason
      - completed_at
      - created_at
      - updated_at
      title: DeletionRequestOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReviewCreate:
      properties:
        prediction_id:
          type: string
          format: uuid
          title: Prediction Id
        review_priority:
          type: string
          title: Review Priority
        review_status:
          type: string
          title: Review Status
        review_action:
          type: string
          title: Review Action
        review_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Review Note
      type: object
      required:
      - prediction_id
      - review_priority
      - review_status
      - review_action
      title: ReviewCreate
    ReviewDetailResponse:
      properties:
        review_id:
          type: string
          format: uuid
          title: Review Id
        prediction_id:
          type: string
          format: uuid
          title: Prediction Id
        reviewer_id:
          type: string
          format: uuid
          title: Reviewer Id
        review_priority:
          type: string
          title: Review Priority
        review_status:
          type: string
          title: Review Status
        review_action:
          type: string
          title: Review Action
        review_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Review Note
        reviewed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Reviewed At
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - review_id
      - prediction_id
      - reviewer_id
      - review_priority
      - review_status
      - review_action
      title: ReviewDetailResponse
    ReviewResponse:
      properties:
        review_id:
          type: string
          format: uuid
          title: Review Id
        prediction_id:
          type: string
          format: uuid
          title: Prediction Id
        saved:
          type: boolean
          title: Saved
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - review_id
      - prediction_id
      - saved
      title: ReviewResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
servers:
- url: http://localhost:8000
  description: Local development server